refactor: extract and refactor shared infrastructure
