+++ gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/shared/components/form/ReviewRenderer.tsx
@@ -1,24 +1,25 @@
+/*
+ * Copyright © 2015 The Gravitee team (http://gravitee.io)
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import { Badge, Button, Card, Switch } from '@gravitee/graphene-core';
-import { BookOpen, ClipboardCheck, Globe, Rocket, Shield } from 'lucide-react';
+import { CircleCheckIcon, FileTextIcon, GlobeIcon, ScrollTextIcon, ShieldIcon } from '@gravitee/graphene-core/icons';
 import { useMemo } from 'react';
 
-import type { ApiCreationState } from '../../domain/apiCreation/models';
-import type { StepConfig } from '../../domain/apiCreation/stepRegistry';
-
-const labelForSecurityType = (type: ApiCreationState['security']['type']): string => {
-    switch (type) {
-        case 'keyless':
-            return 'Keyless (Open)';
-        case 'api-key':
-            return 'API Key';
-        case 'jwt':
-            return 'JWT';
-        case 'oauth2':
-            return 'OAuth 2.0';
-        case 'mtls':
-            return 'mTLS';
-    }
-};
+import type { ApiCreationState } from '../../../features/apis/types/models';
+import { formatSecurityType } from '../../../features/apis/utils/securityFormatters';
+import type { StepConfig } from '../../../utils/stepRegistry';
 
 export type ReviewRendererProps = Readonly<{
     steps: readonly StepConfig[];
@@ -29,29 +30,29 @@ export type ReviewRendererProps = Readonly<{
 
 export function ReviewRenderer({ steps, state, onEditStep, onChangeDeployImmediately }: ReviewRendererProps) {
     const sections = useMemo(() => {
-        const byId = new Map(steps.map((s) => [s.id, s] as const));
+        const byId = new Map(steps.map(step => [step.id, step] as const));
         const ordered = ['api-details', 'configure-proxy', 'secure'] as const;
-        return ordered.map((id) => byId.get(id)).filter(Boolean) as StepConfig[];
+        return ordered.map(id => byId.get(id)).filter(Boolean) as StepConfig[];
     }, [steps]);
 
     return (
         <div className="space-y-4">
             <div className="space-y-1">
                 <div className="flex items-center gap-2">
-                    <ClipboardCheck className="size-4 text-primary" aria-hidden="true" />
+                    <ScrollTextIcon className="size-4 text-primary" aria-hidden="true" />
                     <div className="text-sm font-semibold">Review Configuration</div>
                 </div>
                 <div className="text-xs text-muted-foreground">Click Edit to jump back to any section.</div>
             </div>
 
-            {sections.map((section) => {
+            {sections.map(section => {
                 const SectionIcon =
                     section.id === 'api-details'
-                        ? BookOpen
+                        ? FileTextIcon
                         : section.id === 'configure-proxy'
-                          ? Globe
+                          ? GlobeIcon
                           : section.id === 'secure'
-                            ? Shield
+                            ? ShieldIcon
                             : undefined;
 
                 return (
@@ -104,9 +105,7 @@ export function ReviewRenderer({ steps, state, onEditStep, onChangeDeployImmedia
                                 <div className="sm:col-span-2">
                                     <div className="text-xs text-muted-foreground">Virtual hosts</div>
                                     <div className="font-medium">
-                                        {state.proxy.enableVirtualHosts
-                                            ? `Enabled (${state.proxy.virtualHosts?.length ?? 0})`
-                                            : 'Disabled'}
+                                        {state.proxy.enableVirtualHosts ? `Enabled (${state.proxy.virtualHosts?.length ?? 0})` : 'Disabled'}
                                     </div>
                                 </div>
                             </div>
@@ -117,7 +116,7 @@ export function ReviewRenderer({ steps, state, onEditStep, onChangeDeployImmedia
                                 <div>
                                     <div className="text-xs text-muted-foreground">Authentication</div>
                                     <div className="flex items-center gap-2">
-                                        <Badge variant="secondary">{labelForSecurityType(state.security.type)}</Badge>
+                                        <Badge variant="secondary">{formatSecurityType(state.security.type)}</Badge>
                                     </div>
                                 </div>
                                 {'planName' in state.security ? (
@@ -158,11 +157,13 @@ export function ReviewRenderer({ steps, state, onEditStep, onChangeDeployImmedia
                 <div className="flex items-start justify-between gap-3">
                     <div className="min-w-0">
                         <div className="flex items-center gap-2">
-                            <Rocket className="size-4 text-success" aria-hidden="true" />
+                            <CircleCheckIcon className="size-4 text-success" aria-hidden="true" />
                             <div className="text-sm font-semibold">Deploy and start API immediately</div>
                         </div>
                         <div className="text-xs text-muted-foreground">
-                            {state.deployImmediately ? 'Enabled — proxy will start accepting traffic right away.' : 'Disabled — you can deploy later.'}
+                            {state.deployImmediately
+                                ? 'Enabled — proxy will start accepting traffic right away.'
+                                : 'Disabled — you can deploy later.'}
                         </div>
                     </div>
                     <Switch
@@ -175,4 +176,3 @@ export function ReviewRenderer({ steps, state, onEditStep, onChangeDeployImmedia
         </div>
     );
 }
-
