+++ gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/utils/schema.ts
@@ -1,12 +1,25 @@
-import type { ApiCreationState, SecurityConfig } from './models';
+/*
+ * Copyright © 2015 The Gravitee team (http://gravitee.io)
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import type { ApiCreationState, SecurityConfig } from '../features/apis/types/models';
 
-export type StepId = 'api-details' | 'configure-proxy' | 'secure' | 'review-deploy';
+export type StepId = 'api-details' | 'configure-proxy' | 'secure' | 'review-deploy' | 'essentials';
 
 export type ValidationErrors = Record<string, string>;
 
-export type ValidationResult =
-    | { success: true; errors: {} }
-    | { success: false; errors: ValidationErrors };
+export type ValidationResult = { success: true; errors: {} } | { success: false; errors: ValidationErrors };
 
 const isBlank = (value: unknown) => typeof value !== 'string' || value.trim().length === 0;
 
@@ -30,8 +43,8 @@ export function proxySchema(proxy: ApiCreationState['proxy']): ValidationResult 
     if (proxy.enableVirtualHosts) {
         const any = Array.isArray(proxy.virtualHosts) ? proxy.virtualHosts : [];
         if (any.length === 0) errors['proxy.virtualHosts'] = 'At least one virtual host is required.';
-        if (any.some((v) => isBlank(v.host))) errors['proxy.virtualHosts'] = 'Virtual host is required.';
-        if (any.some((v) => isBlank(v.path))) errors['proxy.virtualHosts'] = 'Context-path is required.';
+        if (any.some(virtualHost => isBlank(virtualHost.host))) errors['proxy.virtualHosts'] = 'Virtual host is required.';
+        if (any.some(virtualHost => isBlank(virtualHost.path))) errors['proxy.virtualHosts'] = 'Context-path is required.';
     } else {
         if (isBlank(proxy.contextPath)) errors['proxy.contextPath'] = 'Context path is required.';
     }
@@ -64,8 +77,25 @@ export function securitySchema(security: ApiCreationState['security']): Validati
     return securityTypeSchema(security);
 }
 
+export function essentialsSchema(data: ApiCreationState): ValidationResult {
+    const d = detailsSchema(data.details);
+    const p = proxySchema(data.proxy);
+    const errors: ValidationErrors = { ...d.errors, ...p.errors };
+
+    if (data.security.type === 'jwt' && isBlank(data.security.resolverParam)) {
+        errors['security.resolverParam'] = 'JWKS URL is required.';
+    }
+    if (data.security.type === 'oauth2' && isBlank(data.security.resource)) {
+        errors['security.resource'] = 'OAuth2 resource is required.';
+    }
+
+    return Object.keys(errors).length ? { success: false, errors } : { success: true, errors: {} };
+}
+
 export function validateStep(stepId: StepId, data: ApiCreationState): ValidationResult {
     switch (stepId) {
+        case 'essentials':
+            return essentialsSchema(data);
         case 'api-details':
             return detailsSchema(data.details);
         case 'configure-proxy':
@@ -73,7 +103,6 @@ export function validateStep(stepId: StepId, data: ApiCreationState): Validation
         case 'secure':
             return securitySchema(data.security);
         case 'review-deploy':
-            return { success: true, errors: {} };
+            return securitySchema(data.security);
     }
 }
-
